requirements.txt
# На сервере без LibreOffice конвертер использует Aspose.Words:
aspose-words>=24.0.0
# Для сборки exe:
pyinstaller>=6.0.0
